requirements.txt
gtkwave
python-pyvcd
python-jinja
python-nmigen
python-nmigen-boards
